Add a virtual shape at the beginning of a shacl path
Ensures that edge-case-01.ttl (one path is a*, another path is b) is
correct. Without this indirection, it may be possible for the engine
to travel a, then propose b as a* can be encoded as an automata
with a single state and a loop transition with symbol a. This change
ensures that the starting node of each inference encoded automata
remains distinct.

Note that due to the current implementation of automata-composer,
a* is translated into an automata with two states, meaning that the
bug does not appear in practice